+++ modules/apple-intelligence/index.ts
@@ -1,5 +1,0 @@
-// Reexport the native module. On web, it will be resolved to AppleIntelligenceModule.web.ts
-// and on native platforms to AppleIntelligenceModule.ts
-export { default } from './src/AppleIntelligenceModule';
-export { default as AppleIntelligenceView } from './src/AppleIntelligenceView';
-export * from  './src/AppleIntelligence.types';