+++ modules/dutch-vocab-ai/src/DutchVocabAI.types.ts
@@ -4,7 +4,7 @@ export type OnLoadEventPayload = {
   url: string;
 };
 
-export type AppleIntelligenceModuleEvents = {
+export type DutchVocabAIModuleEvents = {
   onChange: (params: ChangeEventPayload) => void;
 };
 
@@ -12,7 +12,7 @@ export type ChangeEventPayload = {
   value: string;
 };
 
-export type AppleIntelligenceViewProps = {
+export type DutchVocabAIViewProps = {
   url: string;
   onLoad: (event: { nativeEvent: OnLoadEventPayload }) => void;
   style?: StyleProp<ViewStyle>;